+++ lib/simple_json_persistence.dart
@@ -1,3 +1,7 @@
 library simple_json_persistence;
 
+export 'src/persistence_base.dart';
+export 'src/persistence_noop.dart'
+    if (dart.library.io) 'src/persistence_io.dart'
+    if (dart.library.html) 'src/persistence_html.dart';
 export 'src/simple_json_persistence.dart';