+++ lib/src/persistence_noop.dart
@@ -1,0 +1,4 @@
+import 'package:simple_json_persistence/src/persistence_base.dart';
+
+StoreBackend createStoreBackend([BaseDirectoryBuilder baseDirectoryBuilder]) =>
+    throw UnsupportedError('No implementation found');